+++ PPEventRegistryAPI/PPEventRegistryAPITests/Classes/Transport/PPTransportSpec.swift
@@ -58,28 +58,28 @@ class PPTransportSpec: QuickSpec {
                 waitUntil(timeout: 5) { done in
                     DispatchQueue.global(qos: DispatchQoS.QoSClass.default).async {
                         for _ in 1...100 {
-                            transport.postRequest(controller: .Event, method: .Get, parameters: [:]) { _,_ in }
+                            transport.postRequest(controller: .Event, method: .Get, parameters: [:]) { _ in }
                         }
                     }
                     DispatchQueue.global(qos: DispatchQoS.QoSClass.userInitiated).async {
                         for _ in 1...100 {
-                            transport.postRequest(controller: .Overview, method: .Post, parameters: [:]) { _,_ in }
+                            transport.postRequest(controller: .Overview, method: .Post, parameters: [:]) { _ in }
                         }
                     }
                     DispatchQueue.global(qos: DispatchQoS.QoSClass.background).async {
                         for _ in 1...130 {
-                            transport.postRequest(controller: .Overview, method: .Post, parameters: [:]) { _,_ in }
+                            transport.postRequest(controller: .Overview, method: .Post, parameters: [:]) { _ in }
                         }
                         done()
                     }
                     DispatchQueue.global(qos: DispatchQoS.QoSClass.utility).async {
                         for _ in 1...100 {
-                            transport.postRequest(controller: .Overview, method: .Post, parameters: [:]) { _,_ in }
+                            transport.postRequest(controller: .Overview, method: .Post, parameters: [:]) { _ in }
                         }
                     }
                     DispatchQueue.global(qos: DispatchQoS.QoSClass.userInteractive).async {
                         for _ in 1...100 {
-                            transport.postRequest(controller: .Login, method: .Get, parameters: [:]) { _,_ in }
+                            transport.postRequest(controller: .Login, method: .Get, parameters: [:]) { _ in }
                         }
                     }
                 }
@@ -88,10 +88,8 @@ class PPTransportSpec: QuickSpec {
             it("calls back with 'Response Data is Corrupt' error in case of invalid network response") {
                 PPTransport.stubInvalidResponse()
                 waitUntil { done in
-                    transport.postRequest(controller: .Login, method: .Post, parameters: [:], completionHandler: { response, error in
-                        expect(error?.domain).to(equal("Response Data is Corrupted"))
-                        expect(error?.code).to(equal(0))
-                        expect(response).to(beNil())
+                    transport.postRequest(controller: .Login, method: .Post, parameters: [:], completionHandler: { result in
+                        expect(result.error!.debugDescription).to(equal("CorruptedResponse"))
                         done()
                     })
                 }